+++ src/hooks/useInstancesQuery/useInstancesQuery.js
@@ -7,7 +7,12 @@ import {
 
 // Fetches and returns multiple instances for given instance ids
 const useInstancesQuery = (instances = [], options = {}) => {
-  const { enabled = true, tenantId, ...otherOptions } = options;
+  const {
+    enabled = true,
+    tenantId,
+    include = [],
+    ...otherOptions
+  } = options;
   const [namespace] = useNamespace();
   const ky = useOkapiKy({ tenant: tenantId });
 
@@ -15,10 +20,11 @@ const useInstancesQuery = (instances = [], options = {}) => {
     .map(({ id }) => `id==${id}`)
     .join(' or ');
 
+  const includeProperties = ['identifiers.value', 'identifiers.identifierTypeId', ...include];
   const res = useQuery(
     {
       queryKey: [namespace, 'instances', instanceIdsQuery, tenantId],
-      queryFn: () => ky.get(`search/instances?query=(${instanceIdsQuery})&expandAll=true`).json(),
+      queryFn: () => ky.get(`search/instances?query=(${instanceIdsQuery})&include=${includeProperties.join(',')}`).json(),
       enabled: enabled && Boolean(instanceIdsQuery),
       ...otherOptions,
     },